bug fix midi controller abstract method name, remove output debugger windows, convert pattern generator blending to float32, convert frames to uint8 in luma_key

+++ video_synth/patterns3.py
@@ -206,7 +206,7 @@ class Patterns:
 
         # return pattern
         alpha = self.pattern_alpha.value
-        blended_frame = cv2.addWeighted(frame, 1 - alpha, pattern.astype(np.float32), alpha, 0)
+        blended_frame = cv2.addWeighted(frame.astype(np.float32), 1 - alpha, pattern.astype(np.float32), alpha, 0)
         return blended_frame.astype('uint8')
 
     def _generate_bars(self, pattern: np.ndarray) -> np.ndarray:
